requirements.txt
Django==2.1.2
django-extensions==1.9.9
djangorestframework==3.7.7
eventlet==0.21.0
gunicorn==19.7.1
psycopg2==2.7.3.2
sqlparse==0.2.4